src/main/java/ua/nure/ipz/zoo/entity/Food.java
package ua.nure.ipz.zoo.entity;

import ua.nure.ipz.zoo.util.DomainEntity;

import javax.persistence.Entity;
import javax.persistence.ManyToOne;

@Entity
public class Food extends DomainEntity {
    @ManyToOne
    private Product product;
    private float quantity;

    public Food() {
    }

    public Food(Product product, float quantity) {
        this.product = product;
        this.quantity = quantity;
    }

    public Product getProduct() {
        return product;
    }

    public void setProduct(Product product) {
        this.product = product;
    }

    public float getQuantity() {
        return quantity;
    }

    public void setQuantity(float quantity) {
        this.quantity = quantity;
    }
}
